chore: share common tests between android nad junit tests
