+++ app-shared-tests/src/main/java/com/appunite/loudius/AbsLoginScreenTest.kt
@@ -19,7 +19,6 @@ package com.appunite.loudius
 import android.app.Activity
 import android.app.Instrumentation
 import android.content.Intent
-import android.os.Build
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
@@ -32,34 +31,22 @@ import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
 import androidx.test.espresso.intent.matcher.IntentMatchers.hasData
 import androidx.test.espresso.intent.matcher.IntentMatchers.hasExtra
 import androidx.test.espresso.intent.rule.IntentsRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.appunite.loudius.components.theme.LoudiusTheme
-import com.appunite.loudius.di.GithubHelperModule
 import com.appunite.loudius.ui.login.GithubHelper
 import com.appunite.loudius.ui.login.LoginScreen
 import com.appunite.loudius.util.ScreenshotTestRule
-import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidRule
-import dagger.hilt.android.testing.HiltAndroidTest
-import dagger.hilt.android.testing.HiltTestApplication
-import dagger.hilt.android.testing.UninstallModules
 import io.mockk.every
 import io.mockk.mockk
 import org.hamcrest.Matchers.allOf
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.annotation.Config
 
-@RunWith(AndroidJUnit4::class)
-@UninstallModules(GithubHelperModule::class)
-@Config(sdk = [Build.VERSION_CODES.Q], application = HiltTestApplication::class)
-@HiltAndroidTest
-class LoginScreenTest {
+abstract class AbsLoginScreenTest {
 
     @get:Rule(order = 0)
-    val hiltRule = HiltAndroidRule(this)
+    val hiltRule by lazy { HiltAndroidRule(this) }
 
     @get:Rule(order = 1)
     val composeTestRule = createAndroidComposeRule<TestActivity>()
@@ -76,8 +63,6 @@ class LoginScreenTest {
         hiltRule.inject()
     }
 
-    @BindValue
-    @JvmField
     val githubHelper: GithubHelper = mockk<GithubHelper>().apply {
         every { shouldAskForXiaomiIntent() } returns false
     }
@@ -95,14 +80,13 @@ class LoginScreenTest {
 
         composeTestRule.onNodeWithText("Log in").performClick()
 
-        composeTestRule.runOnIdle {
-            intended(
-                allOf(
-                    hasAction(Intent.ACTION_VIEW),
-                    hasData("https://github.com/login/oauth/authorize?client_id=91131449e417c7e29912&scope=repo"),
-                ),
-            )
-        }
+        composeTestRule.waitForIdle()
+        intended(
+            allOf(
+                hasAction(Intent.ACTION_VIEW),
+                hasData("https://github.com/login/oauth/authorize?client_id=91131449e417c7e29912&scope=repo"),
+            ),
+        )
     }
 
     @Test
@@ -120,14 +104,14 @@ class LoginScreenTest {
         composeTestRule.onNodeWithText("Log in").performClick()
         composeTestRule.onNodeWithText("I've already granted").performClick()
 
-        composeTestRule.runOnIdle {
-            intended(
-                allOf(
-                    hasAction(Intent.ACTION_VIEW),
-                    hasData("https://github.com/login/oauth/authorize?client_id=91131449e417c7e29912&scope=repo"),
-                ),
-            )
-        }
+
+        composeTestRule.waitForIdle()
+        intended(
+            allOf(
+                hasAction(Intent.ACTION_VIEW),
+                hasData("https://github.com/login/oauth/authorize?client_id=91131449e417c7e29912&scope=repo"),
+            ),
+        )
     }
 
     @Test
