+++ app/src/test/java/com/appunite/loudius/UnitLoginScreenTest.kt
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2023 AppUnite S.A.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.appunite.loudius
+
+import android.os.Build
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.appunite.loudius.di.GithubHelperModule
+import dagger.hilt.android.testing.BindValue
+import dagger.hilt.android.testing.HiltAndroidTest
+import dagger.hilt.android.testing.HiltTestApplication
+import dagger.hilt.android.testing.UninstallModules
+import org.junit.runner.RunWith
+import org.robolectric.annotation.Config
+
+@RunWith(AndroidJUnit4::class)
+@Config(sdk = [Build.VERSION_CODES.Q], application = HiltTestApplication::class)
+@HiltAndroidTest
+@UninstallModules(GithubHelperModule::class)
+class UnitLoginScreenTest : AbsLoginScreenTest() {
+
+    @BindValue @JvmField
+    val githubHelperBind = githubHelper
+
+}
